change src 57/58

+++ main.cpp
@@ -77,7 +77,7 @@ int main(int argc, char** argv) {
 
         auto executor = ir::Executor(&program);
         std::cout << program.draw();
-        fprintf(ir::reopen_output_file, "\n%d", executor.run());
+        fprintf(ir::reopen_output_file, "\n%d", (uint8_t) executor.run());
     }
     return 0;
 }
